+++ src/train/util/MaximazingEarlyStop.py
@@ -1,0 +1,11 @@
+from train.util.EarlyStop import EarlyStop
+
+
+class MaximizingEarlyStop(EarlyStop):
+    def __init__(self):
+        super().__init__()
+
+    def validate(self, value)->bool:
+        if self.last_val > value:
+            self.counter += 1
+        return self.stop()
